moe: EP dispatch/combine backends + autotune (TURBO/DeepEP/MORI/UCCL)
- Pluggable EP backends under kernels/moe/_backend (turbo, deep_ep, mori, uccl)
  selected via PRIMUS_TURBO_EP_BACKEND; shared paired dispatch->combine bench +
  sweep_configs autotune harness with per-iter cross-rank barriers.
- deep_ep/TURBO: int64 Buffer/AsymBuffer/SymBuffer offsets (fix int32 overflow on
  large NVL buffers); default EP kernels to the current stream.
- MORI backend (ROCm Mori SHMEM); inter-node mori is not benchmark-safe, so skip
  tuning there and use the base config (live builds one fresh op).
- autotuner: restrict the sweep to a pinned backend; release other backends'
  comm buffers before use (one live EP backend per process).
- uccl: is_available() checks the Buffer API surface, not just the import.
- tests/bench: token-dispatcher tests + _build_tp_groups; deep_ep intranode bench
  supports relative/script imports.

setup.py is left identical to main (mori/uccl are not added as install deps).

+++ csrc/kernels/deep_ep/buffer.cuh
@@ -17,14 +17,14 @@ private:
     uint8_t *ptr;
 
 public:
-    int total_bytes;
+    int64_t total_bytes;
 
     __device__ __forceinline__ Buffer() : ptr(nullptr), total_bytes(0) {}
 
-    __device__ __forceinline__ Buffer(void *&gbl_ptr, int num_elems, int offset = 0) {
-        total_bytes = num_elems * sizeof(dtype_t);
-        ptr         = reinterpret_cast<uint8_t *>(gbl_ptr) + offset * sizeof(dtype_t);
-        gbl_ptr     = reinterpret_cast<uint8_t *>(gbl_ptr) + total_bytes;
+    __device__ __forceinline__ Buffer(void *&gbl_ptr, int64_t num_elems, int64_t offset = 0) {
+        total_bytes = num_elems * static_cast<int64_t>(sizeof(dtype_t));
+        ptr = reinterpret_cast<uint8_t *>(gbl_ptr) + offset * static_cast<int64_t>(sizeof(dtype_t));
+        gbl_ptr = reinterpret_cast<uint8_t *>(gbl_ptr) + total_bytes;
     }
 
     __device__ __forceinline__ Buffer advance_also(void *&gbl_ptr) {
@@ -34,36 +34,36 @@ public:
 
     __device__ __forceinline__ dtype_t *buffer() { return reinterpret_cast<dtype_t *>(ptr); }
 
-    __device__ __forceinline__ dtype_t &operator[](int idx) { return buffer()[idx]; }
+    __device__ __forceinline__ dtype_t &operator[](int64_t idx) { return buffer()[idx]; }
 };
 
 template <typename dtype_t, int kNumRanks = 1> struct AsymBuffer {
 private:
     uint8_t *ptrs[kNumRanks];
-    int      num_bytes;
+    int64_t  num_bytes;
 
 public:
-    int total_bytes;
+    int64_t total_bytes;
 
-    __device__ __forceinline__ AsymBuffer(void *&gbl_ptr, int num_elems, int num_ranks,
-                                          int sm_id = 0, int num_sms = 1, int offset = 0) {
+    __device__ __forceinline__ AsymBuffer(void *&gbl_ptr, int64_t num_elems, int num_ranks,
+                                          int sm_id = 0, int num_sms = 1, int64_t offset = 0) {
         PRIMUS_TURBO_STATIC_CHECK(kNumRanks == 1, "");
-        num_bytes = num_elems * sizeof(dtype_t);
+        num_bytes = num_elems * static_cast<int64_t>(sizeof(dtype_t));
 
-        int per_channel_bytes = num_bytes * num_ranks;
-        total_bytes           = per_channel_bytes * num_sms;
+        int64_t per_channel_bytes = num_bytes * num_ranks;
+        total_bytes               = per_channel_bytes * num_sms;
         ptrs[0] =
             reinterpret_cast<uint8_t *>(gbl_ptr) + per_channel_bytes * sm_id + num_bytes * offset;
         gbl_ptr = reinterpret_cast<uint8_t *>(gbl_ptr) + total_bytes;
     }
 
-    __device__ __forceinline__ AsymBuffer(void **gbl_ptrs, int num_elems, int num_ranks,
-                                          int sm_id = 0, int num_sms = 1, int offset = 0) {
+    __device__ __forceinline__ AsymBuffer(void **gbl_ptrs, int64_t num_elems, int num_ranks,
+                                          int sm_id = 0, int num_sms = 1, int64_t offset = 0) {
         PRIMUS_TURBO_STATIC_CHECK(kNumRanks > 1, "");
-        num_bytes = num_elems * sizeof(dtype_t);
+        num_bytes = num_elems * static_cast<int64_t>(sizeof(dtype_t));
 
-        int per_channel_bytes = num_bytes * num_ranks;
-        total_bytes           = per_channel_bytes * num_sms;
+        int64_t per_channel_bytes = num_bytes * num_ranks;
+        total_bytes               = per_channel_bytes * num_sms;
         for (int i = 0; i < kNumRanks; ++i) {
             ptrs[i] = reinterpret_cast<uint8_t *>(gbl_ptrs[i]) + per_channel_bytes * sm_id +
                       num_bytes * offset;
@@ -71,10 +71,10 @@ public:
         }
     }
 
-    __device__ __forceinline__ void advance(int shift) {
+    __device__ __forceinline__ void advance(int64_t shift) {
 #pragma unroll
         for (int i = 0; i < kNumRanks; ++i)
-            ptrs[i] = ptrs[i] + shift * sizeof(dtype_t);
+            ptrs[i] = ptrs[i] + shift * static_cast<int64_t>(sizeof(dtype_t));
     }
 
     __device__ __forceinline__ AsymBuffer advance_also(void *&gbl_ptr) {
@@ -89,13 +89,13 @@ public:
         return *this;
     }
 
-    __device__ __forceinline__ dtype_t *buffer(int idx = 0) {
+    __device__ __forceinline__ dtype_t *buffer(int64_t idx = 0) {
         PRIMUS_TURBO_STATIC_CHECK(kNumRanks == 1,
                                   "`buffer` is only available for single rank case");
         return reinterpret_cast<dtype_t *>(ptrs[0] + num_bytes * idx);
     }
 
-    __device__ __forceinline__ dtype_t *buffer_by(int rank_idx, int idx = 0) {
+    __device__ __forceinline__ dtype_t *buffer_by(int rank_idx, int64_t idx = 0) {
         PRIMUS_TURBO_STATIC_CHECK(kNumRanks > 1, "`buffer` is only available for single rank case");
         return reinterpret_cast<dtype_t *>(ptrs[rank_idx] + num_bytes * idx);
     }
@@ -106,35 +106,35 @@ private:
     // NOTES: for non-decoupled case, `recv_ptr` is not used
     uint8_t *send_ptr;
     uint8_t *recv_ptr;
-    int      num_bytes;
+    int64_t  num_bytes;
 
 public:
-    int total_bytes;
+    int64_t total_bytes;
 
-    __device__ __forceinline__ SymBuffer(void *&gbl_ptr, int num_elems, int num_ranks,
+    __device__ __forceinline__ SymBuffer(void *&gbl_ptr, int64_t num_elems, int num_ranks,
                                          int sm_id = 0, int num_sms = 1) {
-        num_bytes = num_elems * sizeof(dtype_t);
+        num_bytes = num_elems * static_cast<int64_t>(sizeof(dtype_t));
 
-        int per_channel_bytes = num_bytes * num_ranks;
-        total_bytes           = per_channel_bytes * num_sms * (static_cast<int>(kDecoupled) + 1);
-        send_ptr              = reinterpret_cast<uint8_t *>(gbl_ptr) + per_channel_bytes * sm_id;
-        recv_ptr = reinterpret_cast<uint8_t *>(gbl_ptr) + per_channel_bytes * (sm_id + num_sms);
-        gbl_ptr  = reinterpret_cast<uint8_t *>(gbl_ptr) + total_bytes;
+        int64_t per_channel_bytes = num_bytes * num_ranks;
+        total_bytes = per_channel_bytes * num_sms * (static_cast<int>(kDecoupled) + 1);
+        send_ptr    = reinterpret_cast<uint8_t *>(gbl_ptr) + per_channel_bytes * sm_id;
+        recv_ptr    = reinterpret_cast<uint8_t *>(gbl_ptr) + per_channel_bytes * (sm_id + num_sms);
+        gbl_ptr     = reinterpret_cast<uint8_t *>(gbl_ptr) + total_bytes;
     }
 
-    __device__ __forceinline__ dtype_t *send_buffer(int idx = 0) {
+    __device__ __forceinline__ dtype_t *send_buffer(int64_t idx = 0) {
         PRIMUS_TURBO_STATIC_CHECK(kDecoupled,
                                   "`send_buffer` is only available for non-decoupled case");
         return reinterpret_cast<dtype_t *>(send_ptr + num_bytes * idx);
     }
 
-    __device__ __forceinline__ dtype_t *recv_buffer(int idx = 0) {
+    __device__ __forceinline__ dtype_t *recv_buffer(int64_t idx = 0) {
         PRIMUS_TURBO_STATIC_CHECK(kDecoupled,
                                   "`recv_buffer` is only available for non-decoupled case");
         return reinterpret_cast<dtype_t *>(recv_ptr + num_bytes * idx);
     }
 
-    __device__ __forceinline__ dtype_t *buffer(int idx = 0) {
+    __device__ __forceinline__ dtype_t *buffer(int64_t idx = 0) {
         PRIMUS_TURBO_STATIC_CHECK(not kDecoupled, "`buffer` is only available for decoupled case");
         return reinterpret_cast<dtype_t *>(send_ptr + num_bytes * idx);
     }
